+++ src/pages/scroller/client/statement-history/index.tsx
@@ -1,5 +1,6 @@
 import React, { useMemo, useState } from 'react';
 import { ContentLoader, FilterLayout, SCROLLER_PAGE_LAYOUT_HEADER_HEIGHT, ScrollerPageLayout } from 'components';
+import { FocusNode, FocusTree } from 'components/focus-tree';
 import {
   useAccounts,
   useIsFetchedData,
@@ -11,8 +12,10 @@ import {
 import { useMetricPageListener } from 'hooks/metric/use-metric-page-listener';
 import type { IFilterPanel } from 'interfaces';
 import { Table } from 'pages/scroller/client/statement-history/table';
+import FocusLock from 'react-focus-lock';
 import { getDateRangeValidationScheme } from 'schemas';
 import { DEFAULT_PAGINATION, LINE_HEIGHT, TAB_HEIGHT } from 'stream-constants';
+import { COMMON_SCROLLER_NODE, HISTORY_SCROLLER_FILTER_NODE } from 'stream-constants/a11y-nodes';
 import type { ISortSettings } from '@platform/services';
 import { FatalErrorContent, MainLayout } from '@platform/services/client';
 import { Line } from '@platform/ui';
@@ -124,24 +127,30 @@ export const StatementHistoryScrollerPage = () => {
   return (
     <HistoryScrollerContext.Provider value={contextValue}>
       <MainLayout>
-        <ScrollerPageLayout categoryTabs={tabsProps} headerProps={{ ...headerProps }} loading={!dataFetched}>
-          <ContentLoader height={FILTER_HEIGHT} loading={!accountsFetched}>
-            <FilterLayout
-              AdditionalFilter={AdditionalFilter}
-              QuickFilter={QuickFilter}
-              TagsPanel={TagsPanel}
-              additionalFilterFields={ADDITIONAL_FORM_FIELDS}
-              filterFields={fields}
-              filterState={filterPanel}
-              tagsState={tagsPanel}
-              validate={validate(validationSchema)}
-            />
-          </ContentLoader>
-          {!accountsFetched && <Line fill="FAINT" />}
-          <ContentLoader height={tableHeight} loading={!statementsFetched}>
-            <Table />
-          </ContentLoader>
-        </ScrollerPageLayout>
+        <FocusLock>
+          <FocusTree treeId={COMMON_SCROLLER_NODE}>
+            <ScrollerPageLayout categoryTabs={tabsProps} headerProps={{ ...headerProps }} loading={!dataFetched}>
+              <FocusNode hideBorder nodeId={HISTORY_SCROLLER_FILTER_NODE} parentId={COMMON_SCROLLER_NODE}>
+                <ContentLoader height={FILTER_HEIGHT} loading={!accountsFetched}>
+                  <FilterLayout
+                    AdditionalFilter={AdditionalFilter}
+                    QuickFilter={QuickFilter}
+                    TagsPanel={TagsPanel}
+                    additionalFilterFields={ADDITIONAL_FORM_FIELDS}
+                    filterFields={fields}
+                    filterState={filterPanel}
+                    tagsState={tagsPanel}
+                    validate={validate(validationSchema)}
+                  />
+                </ContentLoader>
+              </FocusNode>
+              {!accountsFetched && <Line fill="FAINT" />}
+              <ContentLoader height={tableHeight} loading={!statementsFetched}>
+                <Table />
+              </ContentLoader>
+            </ScrollerPageLayout>
+          </FocusTree>
+        </FocusLock>
       </MainLayout>
     </HistoryScrollerContext.Provider>
   );