+++ src/pages/scroller/client/statement-transaction/constants.ts
@@ -1,0 +1,7 @@
+/** Мап сортировки для запроса на сервер. */
+export const SORTING_MAP = {
+  /** Списания. */
+  outcome: 'amountDebit',
+  /** Поступления. */
+  income: 'amountCredit',
+};