+++ src/pages/scroller/client/statement-transaction/hooks/index.ts
@@ -1,4 +1,3 @@
 export * from './use-scroller-header-props';
 export * from './use-get-counterparties';
-export * from './use-get-transactions-list';
 export * from './use-get-statement-summary-info';