+++ src/pages/scroller/client/statement-transaction/table/styles.scss.d.ts
@@ -11,7 +11,6 @@ declare namespace StylesScssModule {
     'footer': string;
     'footerDropdownActions': string;
     'rowActionButton': string;
-    'totalWrapper': string;
   }
 }