+++ src/pages/scroller/client/statement-turnover/table/views/view-by-no-group.tsx
@@ -6,7 +6,7 @@ import type { IScrollerView } from '../table-body';
  * Компонент просмотра таблицы без группировки.
  */
 export const ViewByNoGroup: React.FC<IScrollerView> = ({ rows, prepareRow }) => (
-  <AccountList withoutBtn prepareRow={prepareRow} rows={rows[0].subRows} />
+  <AccountList withoutBtn groupRowId={'no-group-id'} prepareRow={prepareRow} rows={rows[0].subRows} />
 );
 
 ViewByNoGroup.displayName = 'ViewByNoGroup';