+++ src/platform-copies/services/components/data-table/components/stop-propagation.tsx
@@ -1,4 +1,3 @@
-import type { FC } from 'react';
 import React, { useCallback } from 'react';
 
 import { Box } from '@platform/ui';
@@ -7,7 +6,7 @@ import { Box } from '@platform/ui';
  * Компонент-обертка для предотвращения всплытия событий.
  * Используется в основном для обертки интерактивных компонентов строки таблицы.
  */
-export const StopPropagation: FC = ({ children }) => {
+export const StopPropagation: React.FC = ({ children }) => {
   const handleClick = useCallback((event: React.SyntheticEvent) => event.stopPropagation(), []);
 
   return <Box onClick={handleClick}>{children}</Box>;